requirements.txt
graphrag
Faker
sentence-transformers 
faiss-cpu
transformers
torch
openai
python-dotenv
numpy